requirements.txt
prompt-toolkit
packaging
python-ls
python-mpv
colr
youtube-dl
youtube-search
clint
pyyaml
requests
pypickle
screen-brightness-control
pathlib
typing
pynput
tabulate
psutil
gputil
webcolors
instaloader